encode with 'unicode-escape'
took me a while to figure out what was wrong, but the presence of an emoji will cause the lxml parser to flip out and throw a "Document is empty" and not parse the response at all. If this happens on the first fetch, the script wont really work. Thanks for this!